refactor: update vars functions to be convert semantics
